+++ github/provider.go
@@ -70,6 +70,7 @@ func Provider() terraform.ResourceProvider {
 		DataSourcesMap: map[string]*schema.Resource{
 			"github_collaborators": dataSourceGithubCollaborators(),
 			"github_ip_ranges":     dataSourceGithubIpRanges(),
+			"github_release":       dataSourceGithubRelease(),
 			"github_repositories":  dataSourceGithubRepositories(),
 			"github_repository":    dataSourceGithubRepository(),
 			"github_team":          dataSourceGithubTeam(),